Tweaks package structures
