+++ helper/runner.go
@@ -6,7 +6,7 @@ import (
 	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/gohcl"
-	"github.com/terraform-linters/tflint-plugin-sdk/terraform"
+	"github.com/terraform-linters/tflint-plugin-sdk/terraform/addrs"
 	"github.com/terraform-linters/tflint-plugin-sdk/terraform/configs"
 	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
 	"github.com/zclconf/go-cty/cty/gocty"
@@ -105,7 +105,7 @@ func (r *Runner) WalkResourceBlocks(resourceType, blockType string, walker func(
 }
 
 // WalkResources visits all specified resources from Files.
-func (r *Runner) WalkResources(resourceType string, walker func(*terraform.Resource) error) error {
+func (r *Runner) WalkResources(resourceType string, walker func(*configs.Resource) error) error {
 	for _, file := range r.Files {
 		resources, _, diags := file.Body.PartialContent(&hcl.BodySchema{
 			Blocks: []hcl.BlockHeaderSchema{
@@ -140,7 +140,7 @@ func (r *Runner) WalkResources(resourceType string, walker func(*terraform.Resou
 }
 
 // WalkModuleCalls visits all module calls from Files.
-func (r *Runner) WalkModuleCalls(walker func(*terraform.ModuleCall) error) error {
+func (r *Runner) WalkModuleCalls(walker func(*configs.ModuleCall) error) error {
 	for _, file := range r.Files {
 		calls, _, diags := file.Body.PartialContent(&hcl.BodySchema{
 			Blocks: []hcl.BlockHeaderSchema{
@@ -171,7 +171,7 @@ func (r *Runner) WalkModuleCalls(walker func(*terraform.ModuleCall) error) error
 }
 
 // Backend returns the terraform backend configuration.
-func (r *Runner) Backend() (*terraform.Backend, error) {
+func (r *Runner) Backend() (*configs.Backend, error) {
 	for _, file := range r.Files {
 		tfcfg, _, diags := file.Body.PartialContent(&hcl.BodySchema{
 			Blocks: []hcl.BlockHeaderSchema{
@@ -193,7 +193,7 @@ func (r *Runner) Backend() (*terraform.Backend, error) {
 			}
 
 			for _, backendBlock := range backendCfg.Blocks {
-				return &terraform.Backend{
+				return &configs.Backend{
 					Type:      backendBlock.Labels[0],
 					TypeRange: backendBlock.LabelRanges[0],
 					Config:    backendBlock.Body,
@@ -229,7 +229,7 @@ func (r *Runner) Config() (*configs.Config, error) {
 				for _, block := range content.Blocks {
 					switch block.Type {
 					case "backend":
-						config.Module.Backend = &terraform.Backend{
+						config.Module.Backend = &configs.Backend{
 							Type:      block.Labels[0],
 							TypeRange: block.LabelRanges[0],
 							Config:    block.Body,
@@ -316,7 +316,7 @@ func (r *Runner) EnsureNoError(err error, proc func() error) error {
 // without depending on Terraform. Some operations have been omitted for ease of implementation.
 // As such, it is expected to parse the minimal code needed for testing.
 // https://github.com/hashicorp/terraform/blob/v0.12.26/configs/resource.go#L78-L288
-func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Diagnostics) {
+func simpleDecodeResouceBlock(resource *hcl.Block) (*configs.Resource, hcl.Diagnostics) {
 	content, resourceRemain, diags := resource.Body.PartialContent(&hcl.BodySchema{
 		Attributes: []hcl.AttributeSchema{
 			{
@@ -352,7 +352,7 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 		forEach = attr.Expr
 	}
 
-	var ref *terraform.ProviderConfigRef
+	var ref *configs.ProviderConfigRef
 	if attr, exists := content.Attributes["provider"]; exists {
 		ref, diags = decodeProviderConfigRef(attr.Expr)
 		if diags.HasErrors() {
@@ -360,7 +360,7 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 		}
 	}
 
-	managed := &terraform.ManagedResource{}
+	managed := &configs.ManagedResource{}
 	for _, block := range content.Blocks {
 		switch block.Type {
 		case "lifecycle":
@@ -399,17 +399,17 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 				}
 			}
 		case "connection":
-			managed.Connection = &terraform.Connection{
+			managed.Connection = &configs.Connection{
 				Config:    block.Body,
 				DeclRange: block.DefRange,
 			}
 		case "provisioner":
-			pv := &terraform.Provisioner{
+			pv := &configs.Provisioner{
 				Type:      block.Labels[0],
 				TypeRange: block.LabelRanges[0],
 				DeclRange: block.DefRange,
-				When:      terraform.ProvisionerWhenCreate,
-				OnFailure: terraform.ProvisionerOnFailureFail,
+				When:      configs.ProvisionerWhenCreate,
+				OnFailure: configs.ProvisionerOnFailureFail,
 			}
 
 			content, config, diags := block.Body.PartialContent(&hcl.BodySchema{
@@ -429,23 +429,23 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 			if attr, exists := content.Attributes["when"]; exists {
 				switch hcl.ExprAsKeyword(attr.Expr) {
 				case "create":
-					pv.When = terraform.ProvisionerWhenCreate
+					pv.When = configs.ProvisionerWhenCreate
 				case "destroy":
-					pv.When = terraform.ProvisionerWhenDestroy
+					pv.When = configs.ProvisionerWhenDestroy
 				}
 			}
 
 			if attr, exists := content.Attributes["on_failure"]; exists {
 				switch hcl.ExprAsKeyword(attr.Expr) {
 				case "continue":
-					pv.OnFailure = terraform.ProvisionerOnFailureContinue
+					pv.OnFailure = configs.ProvisionerOnFailureContinue
 				case "fail":
-					pv.OnFailure = terraform.ProvisionerOnFailureFail
+					pv.OnFailure = configs.ProvisionerOnFailureFail
 				}
 			}
 
 			for _, block := range content.Blocks {
-				pv.Connection = &terraform.Connection{
+				pv.Connection = &configs.Connection{
 					Config:    block.Body,
 					DeclRange: block.DefRange,
 				}
@@ -455,8 +455,8 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 		}
 	}
 
-	return &terraform.Resource{
-		Mode:    terraform.ManagedResourceMode,
+	return &configs.Resource{
+		Mode:    addrs.ManagedResourceMode,
 		Name:    resource.Labels[1],
 		Type:    resource.Labels[0],
 		Config:  resourceRemain,
@@ -472,7 +472,7 @@ func simpleDecodeResouceBlock(resource *hcl.Block) (*terraform.Resource, hcl.Dia
 	}, nil
 }
 
-func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.Diagnostics) {
+func simpleDecodeModuleCallBlock(block *hcl.Block) (*configs.ModuleCall, hcl.Diagnostics) {
 	content, remain, diags := block.Body.PartialContent(&hcl.BodySchema{
 		Attributes: []hcl.AttributeSchema{
 			{Name: "source", Required: true},
@@ -493,7 +493,7 @@ func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.D
 		sourceAddrRange = attr.Expr.Range()
 	}
 
-	providers := []terraform.PassedProviderConfig{}
+	providers := []configs.PassedProviderConfig{}
 	if attr, exists := content.Attributes["providers"]; exists {
 		pairs, diags := hcl.ExprMap(attr.Expr)
 		if diags.HasErrors() {
@@ -511,7 +511,7 @@ func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.D
 				return nil, diags
 			}
 
-			providers = append(providers, terraform.PassedProviderConfig{
+			providers = append(providers, configs.PassedProviderConfig{
 				InChild:  key,
 				InParent: value,
 			})
@@ -537,7 +537,7 @@ func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.D
 		}
 	}
 
-	return &terraform.ModuleCall{
+	return &configs.ModuleCall{
 		Name: block.Labels[0],
 
 		SourceAddr:      sourceAddr,
@@ -546,7 +546,7 @@ func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.D
 
 		Config: remain,
 
-		Version: terraform.VersionConstraint{
+		Version: configs.VersionConstraint{
 			Required:  versionRequired,
 			DeclRange: versionRange,
 		},
@@ -557,13 +557,13 @@ func simpleDecodeModuleCallBlock(block *hcl.Block) (*terraform.ModuleCall, hcl.D
 	}, nil
 }
 
-func decodeProviderConfigRef(expr hcl.Expression) (*terraform.ProviderConfigRef, hcl.Diagnostics) {
+func decodeProviderConfigRef(expr hcl.Expression) (*configs.ProviderConfigRef, hcl.Diagnostics) {
 	traversal, diags := hcl.AbsTraversalForExpr(expr)
 	if diags.HasErrors() {
 		return nil, diags
 	}
 
-	ref := &terraform.ProviderConfigRef{
+	ref := &configs.ProviderConfigRef{
 		Name:      traversal.RootName(),
 		NameRange: traversal[0].SourceRange(),
 	}
